requirements.txt
# Python dependencies
fastapi
uvicorn
kafka-python
neo4j
pydantic
python-dotenv
pandas
huggingface_hub